BackendAPI/pyproject.toml
[project]
name = "backendapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Bruno Haag",email = "proarkes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "fastapi[standard] (>=0.136.0,<0.137.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "sqlalchemy (>=2.0.49,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-cov (>=7.1.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "ruff (>=0.15.11,<0.16.0)"
]


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'


[tool.taskipy.tasks]
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
run = 'fastapi dev BackendAPI/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=BackendAPI -vv'
post_test = 'coverage html'